+++ Makefile
@@ -1,6 +1,16 @@
+.PHONY: fmt lint clean.test test test.clean
+
 CVPKG=go list ./... | grep -v mocks | grep -v internal/
 GO_TEST=go test `$(CVPKG)` -race
 COVERAGE_FILE="coverage.out"
+
+all: fmt lint test install
+
+fmt:
+	go fmt ./...
+
+lint:
+	golangci-lint run --fix ./...
 
 clean.test:
 	go clean --testcache
@@ -12,3 +22,6 @@ test.clean: clean.test test
 
 test.coverage:
 	$(GO_TEST) -covermode=atomic -coverprofile=$(COVERAGE_FILE)
+
+install:
+	go install ./cmd/go-factory-lint
